Move default behavior into library
Make code reuseable and reduce complexity of minimal example (defaultoperator).
Make leader election optional by making it not the default implementation, but a optional class.
