+++ java/operator/org.eclipse.theia.cloud.operator/src/main/java/org/eclipse/theia/cloud/operator/TheiaCloudOperatorLauncher.java
@@ -16,13 +16,43 @@
  ********************************************************************************/
 package org.eclipse.theia.cloud.operator;
 
+import static org.eclipse.theia.cloud.common.util.LogMessageUtil.formatLogMessage;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.eclipse.theia.cloud.operator.di.AbstractTheiaCloudOperatorModule;
 
-public interface TheiaCloudOperatorLauncher {
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 
-    abstract void runMain(String[] args) throws InterruptedException;
+import picocli.CommandLine;
 
-    abstract TheiaCloudOperatorArguments createArguments(String[] args);
+public abstract class TheiaCloudOperatorLauncher {
+
+    static final String COR_ID_INIT = "init";
+
+    private static final Logger LOGGER = LogManager.getLogger(TheiaCloudOperatorLauncher.class);
+
+    protected TheiaCloudOperatorArguments args;
+
+    public void runMain(String[] args) throws InterruptedException {
+	this.args = createArguments(args);
+	AbstractTheiaCloudOperatorModule module = createModule(this.args);
+	LOGGER.info(formatLogMessage(COR_ID_INIT, "Using " + module.getClass().getName() + " as DI module"));
+
+	Injector injector = Guice.createInjector(module);
+	TheiaCloudOperator theiaCloud = injector.getInstance(TheiaCloudOperator.class);
+	LOGGER.info(formatLogMessage(COR_ID_INIT, "Launching Theia Cloud Now"));
+	theiaCloud.start();
+    }
+
+    public TheiaCloudOperatorArguments createArguments(String[] args) {
+	TheiaCloudOperatorArguments arguments = new TheiaCloudOperatorArguments();
+	CommandLine commandLine = new CommandLine(arguments).setTrimQuotes(true);
+	commandLine.parseArgs(args);
+	LOGGER.info(formatLogMessage(COR_ID_INIT, "Parsed args: " + arguments));
+	return arguments;
+    }
 
     abstract AbstractTheiaCloudOperatorModule createModule(TheiaCloudOperatorArguments arguments);
